Add user model tests
